# Conflicts:
#	go.mod
#	go.sum
#	internal/push/push_to_client.go

+++ pkg/common/db/s3/s3.go
@@ -24,7 +24,7 @@ import (
 type PartLimit struct {
 	MinPartSize int64 `json:"minPartSize"`
 	MaxPartSize int64 `json:"maxPartSize"`
-	MaxNumSize  int   `json:"maxNumSize"`
+	MaxNumSize  int64 `json:"maxNumSize"`
 }
 
 type InitiateMultipartUploadResult struct {
